added if request.method == 'DELETE'

+++ api/views.py
@@ -106,15 +106,16 @@ def eligible_jurisdiction(request):
 
 @csrf_exempt
 def unsubscribe(request, phone):
-    formatted_phone = "+1-" + phone
-    alerts = Alert.objects.filter(to=formatted_phone)
-    message = ''
-    if not alerts:
-        message = f"There are no reminders set for {phone}."
-    else:    
-        alerts.delete()  
-        message = f"Reminders for {phone} deleted."
-    return JsonResponse({'message': message})     
+    if request.method == 'DELETE':
+        formatted_phone = "+1-" + phone
+        alerts = Alert.objects.filter(to=formatted_phone)
+        message = ''
+        if not alerts:
+            message = f"There are no reminders set for {phone}."
+        else:    
+            alerts.delete()  
+            message = f"Reminders for {phone} deleted."
+        return JsonResponse({'message': message})     
 
 def find_arraignment_or_return_False(events):
     for event in events:
